requirements.txt


# Flask backend
Flask
Flask-Limiter
Werkzeug
flask-cors
gunicorn
python-dotenv

# Video/audio downloading and processing
pillow
youtube-transcript-api
cloudinary
# Web/API tools
requests
gradio-client
google-generativeai
yt-dlp